Make splash view responsible to show the right view controller
