+++ iOS/Source/AppDelegate.swift
@@ -11,15 +11,7 @@ extension AppDelegate: UIApplicationDelegate {
         self.window = UIWindow(frame: UIScreen.main.bounds)
         guard let window = self.window else { fatalError("Window not found") }
 
-        let dayLightModelController = DaylightModelController()
-
-        if dayLightModelController.location == nil {
-            let controller = OnboardingController(nibName: nil, bundle: nil)
-            window.rootViewController = controller
-        } else {
-            let mainController = MainController(withDaylightModelController: dayLightModelController)
-            window.rootViewController = mainController
-        }
+        window.rootViewController = SplashViewController()
 
         window.makeKeyAndVisible()
